requirements.txt
psycopg2-binary==2.9.7
Django==4.2.5
python-decouple==3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
django-cors-headers==4.2.0
social-auth-app-django==5.3.0
django-templated-mail==1.1.1